Move duplicate code into processConcepts
